Move "Kernel\CompatibilityKernel" to "Internal\CompatibilityTestKernel"
